+++ CHANGES.md
@@ -2,6 +2,7 @@
   * Added more pairs for Kraken and Bitstamp
   * Added support for coinmate.io
   * Widget now fades during data downloading for manually triggered refreshes
+  * Added clickable exchange URL to configuration panel
 
 * v1.0.0 (2021-02-13)
   * Initial public release.